declareWinner imported from a utils file and arrow function in Board onClick event extracted
