+++ rh-tictactoe/src/app/screens/Game/index.js
@@ -1,5 +1,7 @@
 import React from 'react';
 
+import * as utils from '../utils';
+
 import Styles from './styles.scss';
 import Board from './components/Board';
 
@@ -17,12 +19,11 @@ class Game extends React.Component {
     const history = this.state.history;
     const current = history[history.length - 1];
     const squares = current.squares.slice();
-
     let status;
-    if (calculateWinner(squares) === null && squares[i] === null) {
+    if (utils.calculateWinner(squares) === null && squares[i] === null) {
       squares[i] = this.state.xIsNext ? 'X' : 'O';
-      if (calculateWinner(squares)) {
-        status = 'Winner: ' + calculateWinner(squares);
+      if (utils.calculateWinner(squares)) {
+        status = 'Winner: ' + utils.calculateWinner(squares);
       } else {
         status = 'Next player: ' + (!this.state.xIsNext ? 'X' : 'O');
       }
@@ -41,7 +42,7 @@ class Game extends React.Component {
         <div className={Styles.gameBoard}>
           <Board
             squares={this.state.history[this.state.history.length - 1].squares}
-            onClick={i => this.handleClick(i)}
+            onClick={this.handleClick}
             status={this.state.status}
           />
         </div>
@@ -55,23 +56,3 @@ class Game extends React.Component {
 }
 
 export default Game;
-
-function calculateWinner(squares) {
-  let winner = null;
-  const lines = [
-    [0, 1, 2], 
-    [3, 4, 5], 
-    [6, 7, 8], 
-    [0, 3, 6], 
-    [1, 4, 7],
-    [2, 5, 8],
-    [0, 4, 8],
-    [2, 4, 6]
-  ];
-
-  for (let i = 0; i < lines.length; i++) {
-    const [a, b, c] = lines[i];
-    if (squares[a] && squares[a] === squares[b] && squares[a] === squares[c]) winner = squares[a];
-  }
-  return winner;
-}
